move git info struct to a new file
