+++ codex-rs/core/src/codex.rs
@@ -652,9 +652,7 @@ async fn submission_loop(
                 // overlapping rollout file. Consider passing RolloutRecorder
                 // from above.
                 let rollout_recorder =
-                    match RolloutRecorder::new(&config, session_id, instructions.clone(), &cwd)
-                        .await
-                    {
+                    match RolloutRecorder::new(&config, session_id, instructions.clone()).await {
                         Ok(r) => Some(r),
                         Err(e) => {
                             warn!("failed to initialise rollout recorder: {e}");
